requirements.txt
streamlit==1.37.1
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
scikit-image==0.24.0
pillow==10.4.0
scikit-learn==1.5.1
matplotlib==3.9.2
plotly==5.24.1
python-dotenv==1.0.1
tensorflow==2.20.0